+++ core/state/state_candidate.go
@@ -149,3 +149,7 @@ func (c *stateCandidates) deepCopy(db *StateDB, onDirty func()) *stateCandidates
 	stateCandidate := newCandidate(db, c.data, onDirty)
 	return stateCandidate
 }
+
+func (c *stateCandidates) GetData() Candidates {
+	return c.data
+}